flask-server/requirements.txt
langchain
openai
python-dotenv
azure-storage-blob
unstructured[all-docs]
azure-identity
azure-search-documents
langchain-community
Flask
langchain-openai
flask-cors
pymongo
pytz
python-docx
PyMuPDF
gunicorn